bugfix/AB#23188 CHEFs event service
