+++ app/src/forms/form/service.js
@@ -2,8 +2,7 @@ const Problem = require('api-problem');
 const { ref } = require('objection');
 const { v4: uuidv4 } = require('uuid');
 const { EmailTypes, SubscriptionEvent } = require('../common/constants');
-const axios = require('axios');
-const log = require('../../components/log')(module.filename);
+const eventService = require('../event/eventService');
 const moment = require('moment');
 const {
   FileStorage,
@@ -383,21 +382,7 @@ const service = {
       });
 
       await trx.commit();
-
-      const { subscribe } = await service.readForm(formId);
-      if (subscribe && subscribe.enabled) {
-        const subscribeConfig = await service.readFormSubscriptionDetails(formId);
-        const config = Object.assign({}, subscribe, subscribeConfig);
-        const formVersion = new FormVersion();
-        formVersion.id = formVersionId;
-        formVersion.formId = formId;
-
-        if (publish) {
-          service.postSubscriptionEvent(config, formVersion, null, SubscriptionEvent.FORM_PUBLISHED);
-        } else {
-          service.postSubscriptionEvent(config, formVersion, null, SubscriptionEvent.FORM_UNPUBLISHED);
-        }
-      }
+      eventService.publishFormEvent(formId, formVersionId, publish);
 
       // return the published form/version...
       return await service.readPublishedForm(formId);
@@ -445,7 +430,7 @@ const service = {
     let trx;
     try {
       const formVersion = await service.readVersion(formVersionId);
-      const { identityProviders, subscribe } = await service.readForm(formVersion.formId);
+      const { identityProviders } = await service.readForm(formVersion.formId);
 
       trx = await FormSubmission.startTransaction();
 
@@ -498,12 +483,9 @@ const service = {
         };
 
         await FormSubmissionStatus.query(trx).insert(stObj);
-        if (subscribe && subscribe.enabled) {
-          const subscribeConfig = await service.readFormSubscriptionDetails(formVersion.formId);
-          const config = Object.assign({}, subscribe, subscribeConfig);
-          service.postSubscriptionEvent(config, formVersion, submissionId, SubscriptionEvent.FORM_SUBMITTED);
-        }
       }
+
+      eventService.formSubmissionEventReceived(formVersion.formId, formVersion.id, submissionId, data);
 
       // does this submission contain any file uploads?
       // if so, we need to update the file storage records.
@@ -794,35 +776,6 @@ const service = {
     let item = result.length > 0 ? result[0] : null;
     let imageUrl = item !== null ? 'data:' + item.imageType + ';' + 'base64' + ',' + item.image : '';
     return { url: imageUrl };
-  },
-
-  postSubscriptionEvent: async (subscribe, formVersion, submissionId, subscriptionEvent) => {
-    try {
-      // Check if there are endpoints subscribed for form submission event
-      if (subscribe && subscribe.endpointUrl) {
-        const axiosOptions = { timeout: 10000 };
-        const axiosInstance = axios.create(axiosOptions);
-        const jsonData = { formId: formVersion.formId, formVersion: formVersion.id, subscriptionEvent: subscriptionEvent };
-        if (submissionId != null) {
-          jsonData['submissionId'] = submissionId;
-        }
-
-        axiosInstance.interceptors.request.use(
-          (cfg) => {
-            cfg.headers = { [subscribe.key]: `${subscribe.endpointToken}` };
-            return Promise.resolve(cfg);
-          },
-          (error) => {
-            return Promise.reject(error);
-          }
-        );
-        axiosInstance.post(subscribe.endpointUrl, jsonData);
-      }
-    } catch (err) {
-      log.error(err.message, err, {
-        function: 'postSubscriptionEvent',
-      });
-    }
   },
 
   /**
