+++ app/src/forms/submission/service.js
@@ -1,8 +1,9 @@
 const { v4: uuidv4 } = require('uuid');
-
 const { Statuses } = require('../common/constants');
 const { Form, FormVersion, FormSubmission, FormSubmissionStatus, Note, SubmissionAudit, SubmissionMetadata } = require('../common/models');
+const log = require('../../components/log')(module.filename);
 const emailService = require('../email/emailService');
+const eventService = require('../event/eventService');
 const formService = require('../form/service');
 const permissionService = require('../permission/service');
 
@@ -60,6 +61,8 @@ const service = {
     try {
       trx = etrx ? etrx : await FormSubmission.startTransaction();
 
+      log.error('made it in to update');
+      log.error(data);
       // If we're restoring a submission
       if (data['deleted'] !== undefined && typeof data.deleted == 'boolean') {
         await FormSubmission.query(trx).patchAndFetchById(formSubmissionId, { deleted: data.deleted, updatedBy: currentUser.usernameIdp });
@@ -72,6 +75,7 @@ const service = {
             // If finalizing submission, send the submission email (quiet fail if anything goes wrong)
             const submissionMetaData = await SubmissionMetadata.query().where('submissionId', formSubmissionId).first();
             emailService.submissionReceived(submissionMetaData.formId, formSubmissionId, data, referrer).catch(() => {});
+            eventService.formSubmissionEventReceived(submissionMetaData.formId, submissionMetaData.formVersionId, formSubmissionId, data);
           }
         } else {
           if (statuses && statuses.length > 0 && (statuses[0].code === Statuses.SUBMITTED || statuses[0].code === Statuses.COMPLETED)) {
